Fix roles and permissions issues

+++ app/Http/Controllers/Admin/RoleController.php
@@ -70,6 +70,10 @@ class RoleController extends Controller
             abort(403);
         }
 
+        if (in_array($role->name, ['superadmin', 'owner'])) {
+            return back()->withError("System defined roles cannot be edited");
+        }
+
         $roleSpecific = $role->permissions()->get();
         $permissions = Permission::where('panel', 'admin')->get();
         return Inertia::render("Admin/Roles/CreateOrEdit", compact('role', 'permissions', 'roleSpecific'));
@@ -82,6 +86,10 @@ class RoleController extends Controller
     {
         if (! Auth::user()->can('edit_roles')) {
             abort(403);
+        }
+
+        if (in_array($role->name, ['superadmin', 'owner'])) {
+            return back()->withError("System defined roles cannot be edited");
         }
 
         $input = $request->validated();
@@ -97,6 +105,10 @@ class RoleController extends Controller
     {
         if (! Auth::user()->can('delete_roles')) {
             abort(403);
+        }
+
+        if (in_array($role->name, ['superadmin', 'owner'])) {
+            return back()->withError("System defined roles cannot be deleted");
         }
 
         $role->delete();
